+++ src/lightly_train/_methods/detcon/detcon_transform.py
@@ -11,6 +11,7 @@ from pydantic import Field
 
 from lightly_train._configs.config import PydanticConfig
 from lightly_train._transforms.transform import (
+    ChannelDropArgs,
     ColorJitterArgs,
     GaussianBlurArgs,
     MethodTransform,
@@ -56,6 +57,7 @@ class DetConSView1TransformArgs(PydanticConfig):
 
 class DetConSTransformArgs(MethodTransformArgs):
     image_size: tuple[int, int] = Field(default=(224, 224), strict=False)
+    channel_drop: ChannelDropArgs | None = None
     random_resize: RandomResizeArgs | None = Field(default_factory=RandomResizeArgs)
     random_flip: RandomFlipArgs | None = Field(default_factory=RandomFlipArgs)
     random_rotation: RandomRotationArgs | None = None
@@ -106,6 +108,7 @@ class DetConBView1TransformArgs(PydanticConfig):
 
 class DetConBTransformArgs(MethodTransformArgs):
     image_size: tuple[int, int] = Field(default=(224, 224), strict=False)
+    channel_drop: ChannelDropArgs | None = None
     random_resize: RandomResizeArgs | None = Field(default_factory=RandomResizeArgs)
     random_flip: RandomFlipArgs | None = Field(default_factory=RandomFlipArgs)
     random_rotation: RandomRotationArgs | None = None
@@ -129,6 +132,7 @@ class DetConSTransform(MethodTransform):
         # and min from where to sample the (quadratic) size of the kernel
         view_transform_0 = ViewTransform(
             ViewTransformArgs(
+                channel_drop=transform_args.channel_drop,
                 random_resized_crop=RandomResizedCropArgs(
                     size=transform_args.image_size,
                     scale=transform_args.random_resize,
@@ -145,6 +149,7 @@ class DetConSTransform(MethodTransform):
 
         view_transform_1 = ViewTransform(
             ViewTransformArgs(
+                channel_drop=transform_args.channel_drop,
                 random_resized_crop=RandomResizedCropArgs(
                     size=transform_args.image_size,
                     scale=transform_args.random_resize,
@@ -175,6 +180,7 @@ class DetConBTransform(MethodTransform):
 
         view_transform_0 = ViewTransform(
             ViewTransformArgs(
+                channel_drop=transform_args.channel_drop,
                 random_resized_crop=RandomResizedCropArgs(
                     size=transform_args.image_size,
                     scale=transform_args.random_resize,
@@ -191,6 +197,7 @@ class DetConBTransform(MethodTransform):
 
         view_transform_1 = ViewTransform(
             ViewTransformArgs(
+                channel_drop=transform_args.channel_drop,
                 random_resized_crop=RandomResizedCropArgs(
                     size=transform_args.image_size,
                     scale=transform_args.random_resize,