+++ src/lightly_train/_methods/dinov2/dinov2.py
@@ -17,7 +17,6 @@ import torch
 from lightly.loss import (
     KoLeoLoss,
 )  # we use LightlySSL's KoLeoLoss for better numerical stability
-from lightly.models.modules.heads import DINOProjectionHead
 from lightly.models.utils import update_momentum
 from lightly.utils.optim import update_param_groups
 from lightly.utils.scheduler import CosineWarmupScheduler, cosine_schedule
@@ -28,6 +27,7 @@ from torch.nn import Module
 from torch.optim.optimizer import Optimizer
 
 from lightly_train._configs.validate import no_auto
+from lightly_train._methods.dinov2.dinov2_head import DINOv2ProjectionHead
 from lightly_train._methods.dinov2.dinov2_loss import (
     DINOLoss,
     IBOTPatchLoss,
@@ -81,8 +81,12 @@ class DINOv2Args(MethodArgs):
     output_dim: int = 65536  # 65536/131072 for fast/long setup in original DINOv2
     batch_norm: bool = False
     student_freeze_last_layer_epochs: int = 1
-    norm_last_layer: bool = False
-    # NOTE: head_n_layers is 3 for all heads so we use LightlySSL's DINO head
+
+    # freeze student backbone
+    # Useful when starting from DINOv2 pretrained weights. This allows the projection
+    # head to be trained while the backbone is frozen. This is important because the
+    # DINOv2 pretrained weights do not include the projection head.
+    student_freeze_backbone_epochs: int = 0
 
     # loss
     dino_loss_weight: float = 1.0
@@ -161,7 +165,7 @@ class DINOv2AdamWViTArgs(AdamWArgs):
 
 class DINOv2Head(Module):
     def __init__(
-        self, dino_head: DINOProjectionHead, ibot_head: DINOProjectionHead
+        self, dino_head: DINOv2ProjectionHead, ibot_head: DINOv2ProjectionHead
     ) -> None:
         super().__init__()
         self.dino_head = dino_head
@@ -199,35 +203,29 @@ class DINOv2(Method):
 
         # Create teacher and student heads
         dino_head = partial(
-            DINOProjectionHead,
-            input_dim=model.embed_dim,
+            DINOv2ProjectionHead,
+            in_dim=model.embed_dim,
             hidden_dim=method_args.hidden_dim,
             bottleneck_dim=method_args.dino_bottleneck_dim,
-            output_dim=method_args.output_dim,
-            batch_norm=method_args.batch_norm,
-            norm_last_layer=method_args.norm_last_layer,
+            out_dim=method_args.output_dim,
+            use_bn=method_args.batch_norm,
         )
         teacher_dino_head = dino_head()
-        student_dino_head = dino_head(
-            freeze_last_layer=method_args.student_freeze_last_layer_epochs
-        )
+        student_dino_head = dino_head()
 
         ibot_head = partial(
-            DINOProjectionHead,
-            input_dim=model.embed_dim,
+            DINOv2ProjectionHead,
+            in_dim=model.embed_dim,
             hidden_dim=method_args.hidden_dim,
-            bottleneck_dim=method_args.ibot_bottleneck_dim,
-            output_dim=method_args.output_dim,
-            batch_norm=method_args.batch_norm,
-            norm_last_layer=method_args.norm_last_layer,
+            bottleneck_dim=method_args.dino_bottleneck_dim,
+            out_dim=method_args.output_dim,
+            use_bn=method_args.batch_norm,
         )
         teacher_ibot_head = teacher_dino_head
         student_ibot_head = student_dino_head
         if method_args.ibot_separate_head:
             teacher_ibot_head = ibot_head()
-            student_ibot_head = ibot_head(
-                freeze_last_layer=method_args.student_freeze_last_layer_epochs
-            )
+            student_ibot_head = ibot_head()
 
         self.teacher_head = DINOv2Head(
             dino_head=teacher_dino_head, ibot_head=teacher_ibot_head
@@ -238,6 +236,9 @@ class DINOv2(Method):
         freeze_eval_module(self.teacher_head)
 
         # Losses
+        # TODO(Jonas 06/25): make two loss versions one for centering softmax and one for sinkhorn knopp,
+        # so we could instantiate the corresponding one and remove logic form the train loop
+        # LightlySSL solution: https://github.com/lightly-ai/lightly/blob/90ca6abf4cbd34df6e0b58f675d92dc194883602/lightly/models/modules/center.py#L1
         self.dino_loss = DINOLoss(
             out_dim=method_args.output_dim,
             student_temp=method_args.student_temp,
@@ -276,6 +277,7 @@ class DINOv2(Method):
         )  # G * [B, C, H, W] -> [G*B, C, H, W]
 
         # Masking
+        # TODO(Jonas 06/25): put the masking into a separate method
         n_crops = global_views.shape[0]  # G*B
         batch_size = n_crops // n_global_crops
         h = global_views.shape[2] // self._patch_size
@@ -306,6 +308,8 @@ class DINOv2(Method):
         n_masked_patches = mask_indices_list.shape[0]
 
         # Process global views through teacher and student networks
+        # TODO(Jonas 06/25): kwargs
+        # TODO(Jonas 06/25): consider to move all the forwards into a single forward
         teacher_cls_tokens_centered, teacher_masked_patch_tokens_centered = (
             self._forward_teacher(
                 global_views,
@@ -315,14 +319,17 @@ class DINOv2(Method):
                 teacher_temp,
             )  # [G, B, D], [M, D]
         )
-        student_cls_tokens_global, student_masked_patch_tokens_global = (
-            self._forward_student_global(
-                x=global_views,
-                masks=collated_masks,
-                mask_indices_list=mask_indices_list,
-            )  # [G*B, D], [M, D]
-        )
+        (
+            student_cls_tokens_global,
+            student_cls_tokens_global_before_head,
+            student_masked_patch_tokens_global,
+        ) = self._forward_student_global(
+            x=global_views,
+            masks=collated_masks,
+            mask_indices_list=mask_indices_list,
+        )  # [G*B, D], [M, D]
 
+        # TODO(Jonas 06/25): clarify if we actually need this list variant --> simplify interface
         # Compute the DINO loss
         dino_global_loss = (
             self.dino_loss.forward(
@@ -337,6 +344,7 @@ class DINOv2(Method):
 
         # Process local views through student network if they exist
         dino_local_loss = torch.tensor(0.0)
+        # TODO(Jonas 06/25): since n_local_crops is known on instantiation, we could avoid the check and instead instantiate a get_local_views depending on the attribute, similar the forward local could be instantiated like that
         if n_local_crops > 0:
             local_views = torch.cat(
                 views[n_global_crops:]
@@ -345,6 +353,7 @@ class DINOv2(Method):
                 local_views
             )  # [L*B, D]
 
+            # TODO(Jonas 06/25): ideally move everything to tensor only no list
             dino_local_loss = (
                 self.dino_loss.forward(
                     student_output_list=student_cls_tokens_local.chunk(
@@ -365,7 +374,8 @@ class DINOv2(Method):
         )
 
         koleo_loss = sum(
-            self.koleo_loss(token) for token in student_cls_tokens_global.chunk(2)
+            self.koleo_loss(token)
+            for token in student_cls_tokens_global_before_head.chunk(2)
         )  # [G, B, D], only use global views
 
         loss = (
@@ -410,6 +420,7 @@ class DINOv2(Method):
 
         # process the masked patch tokens
         patch_tokens = tokens["features"]  # [G*B, C, H/p, W/p]
+        # TODO(Jonas 06/25): why not flattening the patch tokens here all in one go?
         patch_tokens = patch_tokens.flatten(2).permute(0, 2, 1)  # [G*B, H/p*W/p, C]
 
         masked_patch_tokens = torch.index_select(
@@ -422,19 +433,24 @@ class DINOv2(Method):
         )  # [M, D]
 
         # centering
+        # TODO(Jonas 06/25): instantiate the centering method in the loss and remove the logic from here
         if self.method_args.center_method == "softmax":
+            # TODO(Jonas 06/25): reshape the return inside the loss
             cls_tokens_centered = self.dino_loss.softmax_center_teacher(
                 cls_tokens_after_dino, teacher_temp=teacher_temp
             ).view(2, -1, *cls_tokens_after_dino.shape[1:])  # [G, B, D]
             self.dino_loss.update_center(cls_tokens_after_dino)
 
+            # TODO(Jonas 06/25): change the code inside the loss to avoid the unsqueeze
             masked_patch_tokens_after_ibot = masked_patch_tokens_after_ibot.unsqueeze(0)
             masked_patch_tokens_centered = self.ibot_loss.softmax_center_teacher(
                 masked_patch_tokens_after_ibot,
                 teacher_temp=teacher_temp,
             )  # [M, D]
+            masked_patch_tokens_centered = masked_patch_tokens_centered.squeeze(0)
             self.ibot_loss.update_center(masked_patch_tokens_after_ibot)
         elif self.method_args.center_method == "sinkhorn_knopp":
+            # TODO(Jonas 06/25): reshape the return inside the loss
             cls_tokens_centered = self.dino_loss.sinkhorn_knopp_teacher(
                 cls_tokens_after_dino, teacher_temp=teacher_temp
             ).view(2, -1, *cls_tokens_after_dino.shape[1:])  # [G, B, D]
@@ -442,6 +458,7 @@ class DINOv2(Method):
             masked_patch_tokens_centered = self.ibot_loss.sinkhorn_knopp_teacher(
                 masked_patch_tokens_after_ibot,
                 teacher_temp=teacher_temp,
+                # TODO(Jonas 06/25): move this into the loss if required
                 n_masked_patches_tensor=torch.tensor(
                     [n_masked_patches], dtype=torch.long
                 ).to(device=self.device, non_blocking=True),
@@ -458,7 +475,7 @@ class DINOv2(Method):
         x: Tensor,
         masks: Tensor,
         mask_indices_list: Tensor,
-    ) -> tuple[Tensor, Tensor]:
+    ) -> tuple[Tensor, Tensor, Tensor]:
         wrapped_model: DINOv2ViTModelWrapper = (
             self.student_embedding_model.wrapped_model  # type: ignore[assignment]
         )
@@ -472,6 +489,7 @@ class DINOv2(Method):
 
         # process the patch tokens
         patch_tokens = tokens["features"]  # [G*B, C, H/p, W/p]
+        # TODO(Jonas 06/25): why not flattening the patch tokens here all in one go?
         patch_tokens = patch_tokens.flatten(2).permute(0, 2, 1)  # [G*B, H/p*W/p, C]
 
         masked_patch_tokens = torch.index_select(
@@ -483,7 +501,7 @@ class DINOv2(Method):
             masked_patch_tokens
         )  # [M, D]
 
-        return cls_tokens_after_dino, masked_patch_tokens_after_ibot
+        return cls_tokens_after_dino, cls_tokens, masked_patch_tokens_after_ibot
 
     def _forward_student_local(self, x: Tensor) -> Tensor:
         tokens = self.student_embedding_model.wrapped_model.forward_features(
@@ -491,6 +509,7 @@ class DINOv2(Method):
         )  # input [L*B, C, ...]
 
         # process the cls tokens
+        # TODO(Jonas 06/25): unnecessary assignment, can be removed
         cls_tokens = tokens["cls_token"]  # [L*B, C]
         cls_tokens_after_dino: Tensor = self.student_head.dino_head.forward(
             cls_tokens
@@ -580,9 +599,6 @@ class DINOv2(Method):
         )
 
     def on_before_optimizer_step(self, optimizer: Optimizer, *args: Any) -> None:
-        self.student_head.dino_head.cancel_last_layer_gradients(self.current_epoch)
-        self.student_head.ibot_head.cancel_last_layer_gradients(self.current_epoch)
-
         # Apply weight decay schedule
         weight_decay = cosine_schedule(
             step=self.trainer.global_step,
@@ -593,9 +609,28 @@ class DINOv2(Method):
 
         updates = []
         for group in optimizer.param_groups:
+            update = {}
+            # Apply weight decay schedule
             if group["weight_decay"] != 0.0:
-                updates.append({"name": group["name"], "weight_decay": weight_decay})
+                update["weight_decay"] = weight_decay
 
+            # Optionally freeze student backbone
+            if (
+                self.current_epoch < self.method_args.student_freeze_backbone_epochs
+                and "head" not in group["name"]
+            ):
+                update["lr"] = 0.0
+
+            # Optionally freeze student last layer
+            if (
+                self.current_epoch < self.method_args.student_freeze_last_layer_epochs
+                and "last_layer" in group["name"]
+            ):
+                update["lr"] = 0.0
+
+            if update:
+                update["name"] = group["name"]
+                updates.append(update)
         update_param_groups(optimizer, updates=updates)
 
     def on_train_batch_end(