+++ src/lightly_train/_methods/distillationv2/distillationv2_transform.py
@@ -12,6 +12,7 @@ from __future__ import annotations
 from pydantic import Field
 
 from lightly_train._transforms.transform import (
+    ChannelDropArgs,
     ColorJitterArgs,
     GaussianBlurArgs,
     MethodTransform,
@@ -54,6 +55,7 @@ class DistillationV2GaussianBlurArgs(GaussianBlurArgs):
 
 class DistillationV2TransformArgs(MethodTransformArgs):
     image_size: tuple[int, int] = Field(default=(224, 224), strict=False)
+    channel_drop: ChannelDropArgs | None = None
     random_resize: DistillationV2RandomResizeArgs | None = Field(
         default_factory=DistillationV2RandomResizeArgs
     )
@@ -76,6 +78,7 @@ class DistillationV2Transform(MethodTransform):
 
         self.transform = ViewTransform(
             ViewTransformArgs(
+                channel_drop=transform_args.channel_drop,
                 random_resized_crop=RandomResizedCropArgs(
                     size=transform_args.image_size,
                     scale=transform_args.random_resize,