+++ CHANGELOG.md
@@ -19,6 +19,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.8.1] - 2025-06-23
+
+### Added
+
+- Add `student_freeze_backbone_epochs` option to DINOv2 method to control how many epochs
+  the student backbone is frozen during training. We suggest setting it to 1 when
+  starting from DINOv2 pretrained weights. See the [DINOv2 documentation](https://docs.lightly.ai/train/stable/methods/dinov2.html)
+  for more information.
+- Add channel drop transform.
+- Add option to load multi-channel images with `LIGHTLY_TRAIN_IMAGE_MODE="UNCHANGED"`.
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
 ## [0.8.0] - 2025-06-10
 
 ### Added